pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hephaestus"
version = "0.1.0"
description = "Deterministic metadata governance engine for CI/CD enforcement with OpenMetadata"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Hephaestus Team" },
]

dependencies = [
    "fastapi>=0.110.0,<1.0.0",
    "uvicorn[standard]>=0.27.0,<1.0.0",
    "pydantic>=2.6.0,<3.0.0",
    "pydantic-settings>=2.1.0,<3.0.0",
    "httpx>=0.27.0,<1.0.0",
    "numpy>=1.26.0,<2.0.0",
    "sentence-transformers>=2.3.0,<3.0.0",
    "qdrant-client>=1.7.0,<2.0.0",
    "mcp[server]>=0.1.0",
    "gitpython>=3.1.41,<4.0.0",
    "pyyaml>=6.0.1,<7.0.0",
    "typer>=0.9.0,<1.0.0",
    "rich>=13.7.0,<14.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<1.0.0",
    "ruff>=0.2.0",
]

[project.scripts]
hephaestus-sentinel = "sentinel.main:app"

[tool.hatch.build.targets.wheel]
packages = [
    "config",
    "mcp",
    "sentinel",
    "sync",
    "chronos",
    "api",
    "ingestion",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

